refactor: Organized collection components into sections
