+++ features/collection/components/index.ts
@@ -1,4 +1,2 @@
-export * from './FilterableProductSection';
 export * from './Page';
-export * from './HeroSection';
 export * from './sections';
