+++ features/collection/components/sections/FilterableProductsSection/AppliedFilters.tsx
