+++ features/collection/components/sections/FilterableProductsSection/FilterList/FilterCheckbox.tsx
