+++ features/collection/components/sections/FilterableProductsSection/FilterList/RangeFilter/index.ts
