+++ features/collection/components/sections/FilterableProductsSection/FilterList/virtualAccordion.machine.ts
