+++ features/collection/components/sections/FilterableProductsSection/SearchInput.tsx
