+++ features/collection/components/sections/index.ts
@@ -1,2 +1,6 @@
 export * from './BannerSection';
 export * from './RelatedPostsSection';
+export * from './NewsletterSection';
+export * from './FilterableProductsSection';
+export * from './HeroSection';
+export * from './SubcategoriesSection';
