+++ features/collection/loaders/index.ts
@@ -1,5 +1,5 @@
-import { IFIXIT_API_ORIGIN, STRAPI_ORIGIN } from '@config/env';
-import { Collection, CollectionImage, Post } from '../types';
+import { STRAPI_ORIGIN } from '@config/env';
+import { Collection, CollectionImage } from '../types';
 
 export async function loadCollection(
    handle: string
@@ -30,16 +30,6 @@ async function loadCollectionFromStrapi(
       if (collection == null) {
          return null;
       }
-      const relatedPostSection = collection.sections.find(
-         (section: any) =>
-            section.__typename === 'ComponentCollectionRelatedPosts'
-      );
-      const tags =
-         relatedPostSection?.tags?.split(',').map((tag: any) => tag.trim()) ||
-         [];
-      const relatedPosts = await loadRelatedPosts(
-         [collection.title].concat(tags)
-      );
       return {
          handle: collection.handle,
          title: collection.title,
@@ -54,7 +44,6 @@ async function loadCollectionFromStrapi(
                image: getImageFromStrapiImage(child.image, 'thumbnail'),
             };
          }),
-         relatedPosts,
          sections: collection.sections,
       };
    }
@@ -150,76 +139,3 @@ const collectionByHandleQuery = /* GraphQL */ `
       }
    }
 `;
-
-interface RawPost {
-   ID: number;
-   post_author: string;
-   post_date: string;
-   post_date_gmt: string;
-   post_content: string;
-   post_title: string;
-   post_excerpt: string;
-   permalink?: string;
-   featured_image: RawPostImage;
-   categories: RawPostCategory[];
-}
-
-interface RawPostCategory {
-   term_id: number;
-   name: string;
-   slug: string;
-   term_group: number;
-   term_taxonomy_id: number;
-   taxonomy: string;
-   description: string;
-   parent: number;
-   count: number;
-   filter: string;
-   cat_ID: number;
-   category_count: number;
-   category_description: string;
-   cat_name: string;
-   category_nicename: string;
-   category_parent: number;
-}
-
-interface RawPostImage {
-   thumbnail: string;
-   medium: string;
-   medium_large: string;
-   large: string;
-   '1536x1536': string;
-   '2048x2048': string;
-   'post-header': string;
-   sm: string;
-   md: string;
-   lg: string;
-   'homepage-featured': string;
-   'homepage-small': string;
-   'homepage-medium': string;
-   'rp4wp-thumbnail-post': string;
-}
-
-async function loadRelatedPosts(tags: string[]): Promise<Post[]> {
-   const response = await fetch(
-      `${IFIXIT_API_ORIGIN}/api/2.0/related_posts?data=${encodeURIComponent(
-         JSON.stringify({ tags })
-      )}`
-   );
-   if (response.status >= 200 && response.status < 300) {
-      const rawStories: RawPost[] = await response.json();
-      return rawStories.map<Post>((rawStory) => {
-         return {
-            id: rawStory.ID,
-            title: rawStory.post_title,
-            date: rawStory.post_date,
-            category: rawStory.categories[0]?.name,
-            image: rawStory.featured_image
-               ? { url: rawStory.featured_image.md }
-               : null,
-            permalink: rawStory.permalink || '',
-         };
-      });
-   }
-   throw new Error(`failed with status "${response.statusText}"`);
-}
